todo:fix name not shown when after  added  barn or location element

+++ assets/js/script.js
@@ -1,3 +1,16 @@
+
+formPageModel = function (data) {
+    var self = this;
+    self.data = ko.observableArray(data);
+    self.newDataId = ko.observable();
+    self.getDevName = function (devId) {
+        return ko.utils.arrayFirst(self.data(), function (filter) {
+            return filter.DevId == devId;
+        })
+
+    }
+};
+
 canvasModel = function (data) {
     var self = this;
     self.zoom = ko.observable(data.zoom);
@@ -264,7 +277,7 @@ farmGraphModule.bindJsonElements(function (jsonResponse) {
                     if (route.id() == data.id()) {
                         route.to({
                             guid: $this.guid(),
-                            name: $this.formData().Name
+                            name: $this.formData().DevName
                         });
                     }
                 });
@@ -418,16 +431,10 @@ farmGraphModule.bindJsonElements(function (jsonResponse) {
                         self.activeElement().position().h
                     );
 
-
-
                     var farmItem = { guid: self.activeElement().guid(), position: self.activeElement().position() };
-
                     farmGraphModule.farmDb.setFarmItemSizeAndLocation(farmItem, function (success) {
-
                         if (!success) return;
-
                         self.activeElement().position(positionToSnap);
-
                         //set canvas element position
                         $fg("div[id=" + self.activeElement().guid() + "]").css({
                             width: positionToSnap.w,
@@ -436,94 +443,98 @@ farmGraphModule.bindJsonElements(function (jsonResponse) {
                             top: self.convertToBottomPosition(positionToSnap),
                             left: positionToSnap.x
                         });
-
-
                     });
 
-
+                    self.saveDbData();
                     localStorage.setItem("JSONData", ko.toJSON(self.createdElements()));
                 }
             },
                 viewModel
             );
 
+
+            self.saveDbData = function () {
+
+            }
+
             self.convertToBottomPosition = function (pos) {
                 return (self.canvasProperties().getHeight() - pos.y) - pos.h;
-            },
+            };
+
+            self.setElementPosition = function (event, ui) {
+
+                var zoom = farmGraphModule.elements.drawArea.farmDraw.getZoom();
+                var factor = (1 / zoom) - 1;
+
+                var pos;
+
+                if (ui.size) {
+                    ui.size.width += Math.round((ui.size.width - ui.originalSize.width) * factor);
+                    ui.size.height += Math.round((ui.size.height - ui.originalSize.height) * factor);
+                    ui.position.top += Math.round((ui.position.top - ui.originalPosition.top) * factor);
+                    ui.position.left += Math.round((ui.position.left - ui.originalPosition.left) * factor);
 
 
-                self.setElementPosition = function (event, ui) {
-
-                    var zoom = farmGraphModule.elements.drawArea.farmDraw.getZoom();
-                    var factor = (1 / zoom) - 1;
-
-                    var pos;
-
-                    if (ui.size) {
-                        ui.size.width += Math.round((ui.size.width - ui.originalSize.width) * factor);
-                        ui.size.height += Math.round((ui.size.height - ui.originalSize.height) * factor);
-                        ui.position.top += Math.round((ui.position.top - ui.originalPosition.top) * factor);
-                        ui.position.left += Math.round((ui.position.left - ui.originalPosition.left) * factor);
-
-
-                        pos = {
-                            x: typeof ui.position.left == "number"
-                                ? ui.position.left
-                                : self.getActiveElement().position().x,
-                            y: typeof ui.position.top == "number"
-                                ? ui.position.top
-                                : self.getActiveElement().position().y,
-                            w: typeof ui.size.width == "number"
-                                ? ui.size.width
-                                : self.getActiveElement().position().w,
-                            h: typeof ui.size.height == "number"
-                                ? ui.size.height
-                                : self.getActiveElement().position().h
-                        }
+                    pos = {
+                        x: typeof ui.position.left == "number"
+                            ? ui.position.left
+                            : self.getActiveElement().position().x,
+                        y: typeof ui.position.top == "number"
+                            ? ui.position.top
+                            : self.getActiveElement().position().y,
+                        w: typeof ui.size.width == "number"
+                            ? ui.size.width
+                            : self.getActiveElement().position().w,
+                        h: typeof ui.size.height == "number"
+                            ? ui.size.height
+                            : self.getActiveElement().position().h
                     }
-                    else {
-                        ui.position.top += Math.round((ui.position.top - ui.originalPosition.top) * factor);
-                        ui.position.left += Math.round((ui.position.left - ui.originalPosition.left) * factor);
+                }
+                else {
+                    ui.position.top += Math.round((ui.position.top - ui.originalPosition.top) * factor);
+                    ui.position.left += Math.round((ui.position.left - ui.originalPosition.left) * factor);
 
-                        pos = {
-                            x: typeof ui.position.left == "number"
-                                ? ui.position.left
-                                : self.getActiveElement().position().x,
-                            y: typeof ui.position.top == "number"
-                                ? ui.position.top
-                                : self.getActiveElement().position().y,
-                            w: self.getActiveElement().position().w,
-                            h: self.getActiveElement().position().h
-                        }
-
+                    pos = {
+                        x: typeof ui.position.left == "number"
+                            ? ui.position.left
+                            : self.getActiveElement().position().x,
+                        y: typeof ui.position.top == "number"
+                            ? ui.position.top
+                            : self.getActiveElement().position().y,
+                        w: self.getActiveElement().position().w,
+                        h: self.getActiveElement().position().h
                     }
 
-                    // this section is setting location inputs on Object Information parts
-                    var posToObjectInf = {
-                        left: pos.x,
-                        top: self.convertToBottomPosition(pos),
-                        width: pos.w,
-                        height: pos.h
-                    }
+                }
 
-                    // this section is set size of elements on canvas
-                    pos = farmGraphModule.elements.drawArea.farmDraw.snapToGrid(
-                        pos.x,
-                        pos.y,
-                        pos.w,
-                        pos.h
-                    );
 
-                    posToObjectInf = farmGraphModule.elements.drawArea.farmDraw.snapToGrid(
-                        posToObjectInf.left,
-                        posToObjectInf.top,
-                        posToObjectInf.width,
-                        posToObjectInf.height
-                    )
 
-                    self.getActiveElement().position(posToObjectInf);
-                    return pos;
-                };
+                // this section is setting location inputs on Object Information parts
+                var posToObjectInf = {
+                    left: pos.x,
+                    top: self.convertToBottomPosition(pos),
+                    width: pos.w,
+                    height: pos.h
+                }
+
+                // this section is set size of elements on canvas
+                pos = farmGraphModule.elements.drawArea.farmDraw.snapToGrid(
+                    pos.x,
+                    pos.y,
+                    pos.w,
+                    pos.h
+                );
+
+                posToObjectInf = farmGraphModule.elements.drawArea.farmDraw.snapToGrid(
+                    posToObjectInf.left,
+                    posToObjectInf.top,
+                    posToObjectInf.width,
+                    posToObjectInf.height
+                )
+
+                self.getActiveElement().position(posToObjectInf);
+                return pos;
+            };
 
             self.setEnable = function (acceptable) {
                 var data = self
@@ -565,38 +576,42 @@ farmGraphModule.bindJsonElements(function (jsonResponse) {
                     }
                 };
 
+                deleteelementOnCanvas = function(){
+                    self.createdElements().some(function iter(o, i, a) {
+                        removeAllRoutingRelations(o);
+                        if (o.guid() === activeElement.guid()) {
+                            a.splice(i, 1);
+                            $fg("div[id=" + activeElement.guid() + "]").remove();
+                            self.activeElement(null);
+                        }
+                        var children =
+                            typeof o.children === "function" ? o.children() : o.children;
+                        return children && children.some(iter);
+                    });
+                    // localStorage.setItem("JSONData", ko.toJSON(self.createdElements()));
+                }
+
                 // remove active element
                 if (activeElement) {
 
-                
-                    farmGraphModule.farmDb.RemoveFarmItem(activeElement.guid(),function (success) {
+                    //working this part if new drawed element removing
+                    if (typeof (activeElement.guid()) === 'string'){
+                        deleteelementOnCanvas();
+                        return;
+                    }
+
+                       //working this part if db element removing
+                    farmGraphModule.farmDb.RemoveFarmItem(activeElement.guid(), function (success) {
                         if (!success) return;
-
-                        self.createdElements().some(function iter(o, i, a) {
-                            removeAllRoutingRelations(o);
-                            if (o.guid() === activeElement.guid()) {
-                                a.splice(i, 1);
-                                $fg("div[id=" + activeElement.guid() + "]").remove();
-                                self.activeElement(null);
-                            }
-                            var children =
-                                typeof o.children === "function" ? o.children() : o.children;
-                            return children && children.some(iter);
-                        });
-                        localStorage.setItem("JSONData", ko.toJSON(self.createdElements()));
-                    })
-
+                        deleteelementOnCanvas();
+                    });
 
                 }
             };
 
             self.setTextColor = function () {
-
-
                 var rgb = self.activeElement().color();
-
-
-                if(rgb=='transparent') return "rgb(0,0,0)";
+                if (rgb == 'transparent') return "rgb(0,0,0)";
 
                 var colors = rgb.match(/^rgb\((\d+),\s*(\d+),\s*(\d+)\)$/);
                 var brightness = 5;
